backend/requirements.txt
aioredis==2.0.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
psycopg==3.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.8
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.38.2
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.30.6
websockets==12.0
